exam-platform/backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
httpx>=0.27.0,<1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
psycopg[binary]>=3.1.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
